cloudinary接続機能と二度押し防止機能とコアラ追加編集画面の不具合を修正

+++ src/main/java/com/example/demo/repository/KoalaDao.java
@@ -10,7 +10,7 @@ public interface KoalaDao {
 	
 	int insert(Koala koala);
 
-	Koala findById(Long id);
+	Koala findById(int id);
 	
 	void update(Koala koala);
